feat: support fractional font scale values (0.25 increments)
Changes font.scale and related UI font_scale options from usize to f32,
allowing values like 1.5, 2.25, etc. Zoom now uses 0.25 increments for
smoother scaling. All scale clamping updated to use f32 ranges.

+++ src/app/pane.rs
@@ -111,7 +111,7 @@ impl ShadowPrompt {
 pub struct Pane {
     pub terminal: Terminal,
     pub view: TerminalView,
-    pub scale: usize,
+    pub scale: f32,
 
     pub shell_title: Option<String>,
     pub shell_cwd: Option<String>,
